Improved time precision
- App time and scene time use double instead of float
- App time no longer affected by time scale

+++ Tempus/src/Tempus/Core/Scene.cpp
@@ -7,8 +7,8 @@
 
 void Tempus::Scene::OnUpdate(float DeltaTime)
 {
-    m_SceneTime += DeltaTime;
-
+    m_SceneTime += static_cast<double>(DeltaTime);
+    
     // Update all systems in scene with update enabled
      for (const auto& system : m_Systems)
      {
